[ready event + added get methods for group] feature event (#5)
* [ready event functional] feature event

+++ app/group/group_router.py
@@ -1,19 +1,23 @@
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
+from uuid import UUID
 
 from app.group.crud import GroupCrud
 from app.database import get_session
 from app.users.schemas import GetUser, CreateUserGroupRole
-from app.users.models import GroupRoleEnum
-from app.auth.token_handler import get_current_user_details
+from app.users.models import GroupRoleEnum, RoleEnum
+from app.auth.token_handler import get_current_user_role, get_current_user_details
 from app.group.schemas import CreateGroup, GetGroup
-from app.users.crud import get_user_by_username
+from app.users.crud import UserCrud
 
 group_router = APIRouter(tags=["Group"])
 
 
-@group_router.post("/create_group", response_model=CreateGroup)
-def create_group_endpoint(group: CreateGroup, session: Session = Depends(get_session)):
+@group_router.post("/create_group", response_model=CreateGroup,)
+def create_group_endpoint(group: CreateGroup, session: Session = Depends(get_session),
+                          role: RoleEnum = Depends(get_current_user_role)):
+    if role is not RoleEnum.ADMINISTRATOR:
+        raise HTTPException(status_code=403, detail="You have no rights for this")
     crud = GroupCrud(session)
     db_group = crud.get_group(group)
     if db_group:
@@ -21,8 +25,23 @@ def create_group_endpoint(group: CreateGroup, session: Session = Depends(get_ses
     return crud.create_group(group)
 
 
+@group_router.get("/get_group/{group_id}")
+def get_group_endpoint(group_id: UUID, session: Session = Depends(get_session)):
+    crud = GroupCrud(session)
+    return crud.get_group_by_id(group_id)
+
+
+@group_router.get("/get_all_groups")
+def get_all_groups_endpoint(session: Session = Depends(get_session), skip: int = 0, limit: int = 100):
+    crud = GroupCrud(session)
+    return crud.get_all_groups(skip, limit)
+
+
 @group_router.delete("/delete_group")
-def delete_group_endpoint(group: GetGroup, session: Session = Depends(get_session)):
+def delete_group_endpoint(group: GetGroup, session: Session = Depends(get_session),
+                          role: RoleEnum = Depends(get_current_user_role)):
+    if role is not RoleEnum.ADMINISTRATOR:
+        raise HTTPException(status_code=403, detail="You have no rights for this")
     crud = GroupCrud(session)
     return crud.delete_group(group)
 
@@ -55,8 +74,12 @@ def delete_user_from_group_endpoint(user: GetUser = Depends(get_current_user_det
 
 
 @group_router.put("/assign_new_group_role_to_user")
-def update_group_role_endpoint(username: str, new_role: GroupRoleEnum, session: Session = Depends(get_session)):
-    user = get_user_by_username(username=username, session=session)
+def update_group_role_endpoint(username: str, new_role: GroupRoleEnum, session: Session = Depends(get_session),
+                               role: RoleEnum = Depends(get_current_user_role)):
+    if role is RoleEnum.DEFAULT_USER:
+        raise HTTPException(status_code=403, detail="You have no rights for this")
+    crud = UserCrud(session)
+    user = crud.get_user_by_username(username=username)
     crud = GroupCrud(session)
     user_group_role = crud.get_user_group_role(user)
     if not user_group_role:
